Clarified Language Surrounding Hat Collisions
Made it clearer what has gone wrong when a specified hat cannot be added because one already exists for it in the KBX. Also changed the message associated with this issue in the changelog to be tagged as a warning to be consistent with the message displayed in the console.

Additionally, I'd misconfigured some comparison conditions such that the highest valid EX Fighter ID (0x7F) wasn't actually bein indexed within the KBX file, and thus wasn't accurately noted in logs or checked for collisions. This has now been fixed, such hats should now be properly detected and checked against!